+++ app/layout.tsx
@@ -1,27 +1,17 @@
 "use client";
 
-import React, { Suspense, createContext, useEffect, useState } from "react";
+import React, { Suspense, useEffect, useState } from "react";
 import "./globals.css";
 import posthog from "posthog-js";
-import { PHProvider, PostHogPageview } from "./providers";
+import { GlobalContext, PHProvider, PostHogPageview } from "./providers";
 import { Header } from "@/components/header";
-import * as pako from "pako";
+import pako from "pako";
 
 if (typeof window !== "undefined") {
   posthog.init(process.env.NEXT_PUBLIC_POSTHOG_KEY, {
     api_host: process.env.NEXT_PUBLIC_POSTHOG_HOST,
   });
 }
-
-type GlobalContextType = {
-  setSource: (source: string) => void;
-  source: string;
-};
-
-export const GlobalContext = createContext<GlobalContextType>({
-  source: "",
-  setSource: () => {},
-});
 
 const encode = (source: string) => {
   const input = new TextEncoder().encode(source);
@@ -40,9 +30,10 @@ export default function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
+  const hash = typeof window !== "undefined" ? window.location.hash : "";
   const [source, setSource] = useState<string>(
-    window.location.hash
-      ? decode(window.location.hash)
+    hash
+      ? decode(hash)
       : `title 'Shadow Priest WhackAura'
 load spec: :shadow_priest
 hide_ooc!
@@ -54,13 +45,13 @@ end`
 
   useEffect(() => {
     const base64 = encode(source);
-    window.location.hash = base64;
+    if (window?.location) window.location.hash = base64;
     history.pushState(null, "", `#${base64}`);
   }, [source]);
 
   useEffect(() => {
     const handleHashChange = () => {
-      setSource(decode(window.location.hash));
+      setSource(decode(window?.location?.hash));
     };
     window.addEventListener("hashchange", handleHashChange);
     return () => {
